4.1 - Automatic collection-based sub-layouts.
Collections (the 'posts' folder in this case) will automatically use any layout of the same name. This eliminates the need to define your layout on every page. Here, ruhoh sees the layout 'posts.html' so it will enable this layout by default for the 'posts' collection. You are still free to explicitly define a layout selection inside of a given page.