7 - Stylesheet fingerprinting.
Up until now we have loaded our stylesheet in the stock way. The url to the stylesheet was "/stylesheets/application.css". However, frequently updated websites are faced with the issue of web-browsers storing a cached copy of stylesheets locally. This means that you may push CSS updates but users may not see them since their browser will load the outdated cached version.

The only bullet-proof way around ensuring all users get the most up-to-date stylesheet is to constantly rename the stylesheet something new. This is called "asset fingerprinting" where the name of the file is appended with a hash digest of the contents of the file. The filename only changes if the content changes.

Ruhoh takes care of this for you automatically if you load your stylesheets (and javascripts) via the ruhoh load syntax. Now your urls look like this: "/assets/stylesheets/application-6588954608fec50d7e5c3eefda9b6544.css" and are guaranteed to be unique as their content gets modified.